added pcie errors in pcie_device
