+++ sysfs/pci_device.go
@@ -99,6 +99,61 @@ type PciDevice struct {
 	PowerState    *PciPowerState // /sys/bus/pci/devices/<Location>/power_state
 }
 
+// CorrectableAerCounters contains values from /sys/class/net/<iface>/device/aer_dev_correctable
+// for single interface (iface).
+type CorrectableAerCounters struct {
+	RxErr       uint64
+	BadTLP      uint64
+	BadDLLP     uint64
+	Rollover    uint64
+	Timeout     uint64
+	NonFatalErr uint64
+	CorrIntErr  uint64
+	HeaderOF    uint64
+	TotalErrCor uint64 // TOTAL_ERR_COR from aer_dev_correctable
+}
+
+// UncorrectableAerCounters contains values from /sys/class/net/<iface>/device/aer_dev_[non]fatal
+// for single interface (iface).
+type UncorrectableAerCounters struct {
+	Undefined        uint64
+	DLP              uint64
+	SDES             uint64
+	TLP              uint64
+	FCP              uint64
+	CmpltTO          uint64
+	CmpltAbrt        uint64
+	UnxCmplt         uint64
+	RxOF             uint64
+	MalfTLP          uint64
+	ECRC             uint64
+	UnsupReq         uint64
+	ACSViol          uint64
+	UncorrIntErr     uint64
+	BlockedTLP       uint64
+	AtomicOpBlocked  uint64
+	TLPBlockedErr    uint64
+	PoisonTLPBlocked uint64
+	TotalErrFatal    uint64 // TOTAL_ERR_FATAL from aer_dev_fatal
+	TotalErrNonFatal uint64 // TOTAL_ERR_NONFATAL from aer_dev_nonfatal
+}
+
+// AerCounters contains AER counters from files in /sys/class/net/<iface>/device
+// for single interface (iface).
+type AerCounters struct {
+	Name                     string // Interface name
+	Correctable              CorrectableAerCounters
+	Fatal                    UncorrectableAerCounters
+	NonFatal                 UncorrectableAerCounters
+	RootPortTotalErrCor      *uint64 // aer_rootport_total_err_cor (optional, may not exist)
+	RootPortTotalErrFatal    *uint64 // aer_rootport_total_err_fatal (optional, may not exist)
+	RootPortTotalErrNonFatal *uint64 // aer_rootport_total_err_nonfatal (optional, may not exist)
+}
+
+// AllAerCounters is collection of AER counters for every interface (iface) in /sys/class/net.
+// The map keys are interface (iface) names.
+type AllAerCounters map[string]AerCounters
+
 func (pd PciDevice) Name() string {
 	return pd.Location.String()
 }
@@ -405,4 +460,231 @@ func (fs FS) parsePciDevice(name string) (*PciDevice, error) {
 	}
 
 	return device, nil
+}
+
+// ParseAerCounters scans predefined files in /sys/class/net/<iface>/device
+// directory and gets their contents.
+func ParseAerCounters(devicePath string) (*AerCounters, error) {
+	counters := AerCounters{}
+	err := parseCorrectableAerCounters(devicePath, &counters.Correctable)
+	if err != nil {
+		return nil, err
+	}
+	err = parseUncorrectableAerCounters(devicePath, "fatal", &counters.Fatal)
+	if err != nil {
+		return nil, err
+	}
+	err = parseUncorrectableAerCounters(devicePath, "nonfatal", &counters.NonFatal)
+	if err != nil {
+		return nil, err
+	}
+
+	// Parse root port error files (optional, may not exist)
+	err = parseRootPortAerCounters(devicePath, &counters)
+	if err != nil {
+		return nil, err
+	}
+
+	return &counters, nil
+}
+
+// parseRootPortAerCounters parses root port AER error counters from
+// /sys/class/net/<iface>/device/aer_rootport_total_err_* files.
+func parseRootPortAerCounters(devicePath string, counters *AerCounters) error {
+	deviceDir := filepath.Join(devicePath, "device")
+
+	// Parse aer_rootport_total_err_cor
+	path := filepath.Join(deviceDir, "aer_rootport_total_err_cor")
+	value, err := util.SysReadFile(path)
+	if err != nil {
+		if canIgnoreError(err) {
+			// File doesn't exist or can't be read, skip it
+		} else {
+			return fmt.Errorf("failed to read file %q: %w", path, err)
+		}
+	} else {
+		valueStr := strings.TrimSpace(string(value))
+		if valueStr != "" {
+			v, err := strconv.ParseUint(valueStr, 10, 64)
+			if err != nil {
+				return fmt.Errorf("error parsing aer_rootport_total_err_cor: %w", err)
+			}
+			counters.RootPortTotalErrCor = &v
+		}
+	}
+
+	// Parse aer_rootport_total_err_fatal
+	path = filepath.Join(deviceDir, "aer_rootport_total_err_fatal")
+	value, err = util.SysReadFile(path)
+	if err != nil {
+		if canIgnoreError(err) {
+			// File doesn't exist or can't be read, skip it
+		} else {
+			return fmt.Errorf("failed to read file %q: %w", path, err)
+		}
+	} else {
+		valueStr := strings.TrimSpace(string(value))
+		if valueStr != "" {
+			v, err := strconv.ParseUint(valueStr, 10, 64)
+			if err != nil {
+				return fmt.Errorf("error parsing aer_rootport_total_err_fatal: %w", err)
+			}
+			counters.RootPortTotalErrFatal = &v
+		}
+	}
+
+	// Parse aer_rootport_total_err_nonfatal
+	path = filepath.Join(deviceDir, "aer_rootport_total_err_nonfatal")
+	value, err = util.SysReadFile(path)
+	if err != nil {
+		if canIgnoreError(err) {
+			// File doesn't exist or can't be read, skip it
+		} else {
+			return fmt.Errorf("failed to read file %q: %w", path, err)
+		}
+	} else {
+		valueStr := strings.TrimSpace(string(value))
+		if valueStr != "" {
+			v, err := strconv.ParseUint(valueStr, 10, 64)
+			if err != nil {
+				return fmt.Errorf("error parsing aer_rootport_total_err_nonfatal: %w", err)
+			}
+			counters.RootPortTotalErrNonFatal = &v
+		}
+	}
+
+	return nil
+}
+
+// parseCorrectableAerCounters parses correctable error counters in
+// /sys/class/net/<iface>/device/aer_dev_correctable.
+func parseCorrectableAerCounters(devicePath string, counters *CorrectableAerCounters) error {
+	path := filepath.Join(devicePath, "device", "aer_dev_correctable")
+	value, err := util.SysReadFile(path)
+	if err != nil {
+		if canIgnoreError(err) {
+			return nil
+		}
+		return fmt.Errorf("failed to read file %q: %w", path, err)
+	}
+
+	for line := range strings.SplitSeq(string(value), "\n") {
+		if line == "" {
+			continue
+		}
+		fields := strings.Fields(line)
+		if len(fields) != 2 {
+			return fmt.Errorf("unexpected number of fields: %v", fields)
+		}
+		counterName := fields[0]
+		value, err := strconv.ParseUint(fields[1], 10, 64)
+		if err != nil {
+			return fmt.Errorf("error parsing value for %s: %w", counterName, err)
+		}
+
+		switch counterName {
+		case "RxErr":
+			counters.RxErr = value
+		case "BadTLP":
+			counters.BadTLP = value
+		case "BadDLLP":
+			counters.BadDLLP = value
+		case "Rollover":
+			counters.Rollover = value
+		case "Timeout":
+			counters.Timeout = value
+		case "NonFatalErr":
+			counters.NonFatalErr = value
+		case "CorrIntErr":
+			counters.CorrIntErr = value
+		case "HeaderOF":
+			counters.HeaderOF = value
+		case "TOTAL_ERR_COR":
+			counters.TotalErrCor = value
+		default:
+			continue
+		}
+	}
+
+	return nil
+}
+
+// parseUncorrectableAerCounters parses uncorrectable error counters in
+// /sys/class/net/<iface>/device/aer_dev_[non]fatal.
+func parseUncorrectableAerCounters(devicePath string, counterType string,
+	counters *UncorrectableAerCounters) error {
+	path := filepath.Join(devicePath, "device", "aer_dev_"+counterType)
+	value, err := util.ReadFileNoStat(path)
+	if err != nil {
+		if canIgnoreError(err) {
+			return nil
+		}
+		return fmt.Errorf("failed to read file %q: %w", path, err)
+	}
+
+	for line := range strings.SplitSeq(string(value), "\n") {
+		if line == "" {
+			continue
+		}
+		fields := strings.Fields(line)
+		if len(fields) != 2 {
+			return fmt.Errorf("unexpected number of fields: %v", fields)
+		}
+		counterName := fields[0]
+		value, err := strconv.ParseUint(fields[1], 10, 64)
+		if err != nil {
+			return fmt.Errorf("error parsing value for %s: %w", counterName, err)
+		}
+
+		switch counterName {
+		case "Undefined":
+			counters.Undefined = value
+		case "DLP":
+			counters.DLP = value
+		case "SDES":
+			counters.SDES = value
+		case "TLP":
+			counters.TLP = value
+		case "FCP":
+			counters.FCP = value
+		case "CmpltTO":
+			counters.CmpltTO = value
+		case "CmpltAbrt":
+			counters.CmpltAbrt = value
+		case "UnxCmplt":
+			counters.UnxCmplt = value
+		case "RxOF":
+			counters.RxOF = value
+		case "MalfTLP":
+			counters.MalfTLP = value
+		case "ECRC":
+			counters.ECRC = value
+		case "UnsupReq":
+			counters.UnsupReq = value
+		case "ACSViol":
+			counters.ACSViol = value
+		case "UncorrIntErr":
+			counters.UncorrIntErr = value
+		case "BlockedTLP":
+			counters.BlockedTLP = value
+		case "AtomicOpBlocked":
+			counters.AtomicOpBlocked = value
+		case "TLPBlockedErr":
+			counters.TLPBlockedErr = value
+		case "PoisonTLPBlocked":
+			counters.PoisonTLPBlocked = value
+		case "TOTAL_ERR_FATAL":
+			if counterType == "fatal" {
+				counters.TotalErrFatal = value
+			}
+		case "TOTAL_ERR_NONFATAL":
+			if counterType == "nonfatal" {
+				counters.TotalErrNonFatal = value
+			}
+		default:
+			continue
+		}
+	}
+
+	return nil
 }
